+++ src/clj/rems/auth/oidc.clj
@@ -15,6 +15,7 @@
        "&client_id=" (getx env :oidc-client-id)
        "&redirect_uri=" (getx env :public-url) "oidc-callback"
        "&scope=openid profile email"
+       (getx env :oidc-additional-authorization-parameters)
        #_"&state=STATE")) ; FIXME We could use the state for intelligent redirect. Also check if we need it for CSRF protection as Auth0 docs say.
 
 (defn logout-url []